Move XPDateFormatter to dedicated file
