+++ core/ui/src/commonMain/kotlin/design/Theme.kt
@@ -10,22 +10,7 @@ import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.compositionLocalOf
 import org.jetbrains.compose.ui.tooling.preview.Preview
-import kotlin.time.ExperimentalTime
-import kotlin.time.Instant
-
-@OptIn(ExperimentalTime::class)
-interface XPDateFormatter {
-    fun format(date: Instant): String
-}
-@OptIn(ExperimentalTime::class)
-val SimpleDateFormatter = object : XPDateFormatter {
-    override fun format(date: Instant): String {
-        return date.toString()
-    }
-}
-val LocalDateFormat = compositionLocalOf<XPDateFormatter> { throw IllegalStateException("DateFormat not provided") }
 
 /**
  * Extends the Material3 Theme with an [XPDateFormatter].
